rag-assist/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-assist"
version = "0.1.0"
description = "Production RAG Learning Assistant using AWS Strands SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "RAG Assist Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AWS SDK
    "boto3>=1.35.0",
    "botocore>=1.35.0",

    # AWS Strands Agents SDK
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",

    # Document Parsing
    "pymupdf>=1.24.0",
    "python-pptx>=1.0.0",
    "python-docx>=1.1.0",
    "amazon-textract-textractor>=1.7.0",

    # NLP & Embeddings
    "spacy>=3.7.0",
    "yake>=0.4.8",
    "keybert>=0.8.0",
    "sentence-transformers>=3.0.0",
    "datasketch>=1.6.0",

    # Vector Store
    "opensearch-py>=2.5.0",
    "opensearch-dsl>=2.1.0",

    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=8.5.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "moto[all]>=5.0.0",
    "localstack>=3.7.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "ragas>=0.1.0",
]

test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "moto[all]>=5.0.0",
]

[project.scripts]
rag-assist = "rag_assist.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_assist"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/rag_assist --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
